chop: force double precision

+++ pytoolsMH/math.py
@@ -89,7 +89,7 @@ def smooth_spline(y, x=None, knots=None, degree=3):
 def chop(x0, sig=2):
     x0 = np.atleast_1d(x0)
     zeroIx = np.array(x0)==0
-    x = np.array(x0)
+    x = np.array(x0, dtype='f8')  # must be float for chop to make sense.  Later, support single precision?
     x[zeroIx] = 1  # will unmask below
     nSig = sig-(np.floor(np.log10(x)))-1
     if np.size(x) == 1:
